Replace echo -en with printf
```bash
echo -en "some string to escape"
```

Is not part of the POSIX standard. This creates problems when /usr/bin/sh is symlinked to a strictly POSIX compliant shell (e.g. dash). You can use a tool like checkbashisms to see where the issues lie; which there are because the shebang line claims to be a pure shell script. This can be fixed in two ways.

1. Change the shebang to `#!/usr/bin/bash`.
2. Fix the lines pointed out by checkbashisms to strict POSIX compliance. I choose this option and have submitted this as a pull request.